requirements.txt
cffi==1.15.0
click==7.1.2
cryptography==36.0.2
pyAesCrypt==6.0.0
pycparser==2.21
PyQt5==5.15.6
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
PyQt5-stubs==5.15.2.0
pyqt5-tools==5.15.4.3.2
python-dotenv==0.20.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
